Implement Bram for tennoko bank.
When it first runs, press 1,2 and 3 and it will get a bit further. no idea what the fuck it's supposed to be however, but it seems to want a card inserted somewhere. Might need proper CD support.

Also, someone might need to clean up my mess a bit, I have forgotten how to do bitmasking.