All I wanted to do was clean up the todo against hcastle...
I had a look to see why hcastle was crashing when using the correct refresh rate. It turns out that nBurnSoundLen is not calculated correctly until after DrvInit if a game is not using 60hz. This probably affects a lot of games and possibly might need further verifcation on others to make sure that the sound is working correctly now it's calculated correctly. Some sound cores didn't care about nBurnSoundLen when creating their internal buffers but that ones that do would always have a buffer underrun.

I also think I have found a fix for starwars crashing randonmly on my machine but if someone could verify that the games still work on other machines that would be good ;)

Some notes:
Wingforc - someone should check that the changes I made are correct :)
hcastle - seems ok with the refresh rate now
yiear - might be ok as is now I have renabled the refrest rate. Might not be. Could someone check.
namcos1 - Seems the games that are above 60hz now work with the correct refresh but someone should double check

Remaining refrest problems I know of:
metlclsh - Had a note that it locks up with refrest rate, seems to still crash so I have left it uncommented
M62 - Wasn't sure if the refrest rate was commented out due to sound so have left this