This clause doesn't really make sense
Since the project is reworking its license, it might be a good opportunity to remove this clause, it's pretty vague and i don't think that's something enforcable. If it is read verbatim it might even prevent the project from being usable at all (the project relies on directx, which is a bunch of dlls).