Add some MS hacks with HP indicator.
@dinkc64  As you requested, I added them.
And, could you do me a favor on hacks?
I want to add a hack of KOF 97(kof97t in HBMAME) to FB Neo, and this hack is the most popular hack of KOF97 from GOTVG. However, it is encryted and has protections. I have decrypted it and submitted it to HBMAME. Robbert has patched one protection, but there are more protections left. Can you take a look at it and help remove the protections? 
Thank you in advance. 
If you have any trouble in getting the ROMs I added, please contract me.
Regards.