SDL2: fix inputs from <romname>.ini being ignored (and presets/*.ini too)
Any changes made to inputs would always be ignored and reset to defaults.

Also fixed presets being ignored in SDL builds (for example, presets/cps.ini will be used for any CPS game if <romname>.ini did not exist).

Note about -joy parameter and player 1 using keyboard or joystick:
If <romname>.ini or presets/*.ini exists when a game is launched, whatever is defined for player 1 is what is used, being irrelevant if -joy parameter is used or not.