d_taitof3: Ensure EEPROM data is actually saved
Because the function that should be doing that ... wasn't. The behavior
is commented out, and has been since FBNeo's first commit.

Since I don't know _why_ this is the case, and I don't feel comfortable
changing the code for every single driver, a new function was introduced
to specifically copy the data contained within the EEPROM, that the F3
driver now uses.